podcast

+++ app/Http/Controllers/Admin/BookController.php
@@ -13,6 +13,8 @@ use App\Models\Post;
 use App\Models\PostBook;
 use Illuminate\Http\Request;
 
+use function PHPUnit\Framework\isEmpty;
+
 class BookController extends Controller
 {
     protected $modelName = 'App\Models\Book';
@@ -43,8 +45,10 @@ class BookController extends Controller
     public function destroy($id)
     {
         $postBookId = PostBook::where('book_id', $id)->pluck('id');
-        Post::where('relationable_type', 'App\\Models\\PostBook')->where('relationable_id', $postBookId)->delete();
-        PostBook::where('book_id', $id)->delete();
+        if (count($postBookId) > 0) {
+            Post::where('relationable_type', 'App\\Models\\PostBook')->where('relationable_id', $postBookId)->delete();
+            PostBook::where('book_id', $id)->delete();
+        }
         CRUD::delete($this->modelName, $id);
         return back()->with('success', 'Maglumat üstünlikli pozuldy');
     }
